fastapi-graphql/requirements.txt
aniso8601==7.0.0
boto==2.49.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
elasticsearch==7.11.0
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==2.10
promise==2.3
pydantic==1.7.3
python-dotenv==0.17.0
requests==2.25.1
Rx==1.6.1
six==1.15.0
starlette==0.13.6
urllib3==1.26.3
uuid==1.30
uvicorn==0.13.3
